Refactor Campeonato __str__ method to include year and mata-mata flag
Refactor the __str__ method in the Campeonato model to include the year and whether it is a mata-mata phase or not. This provides more descriptive information when displaying the Campeonato object.

Refactor Time model fields order

Reorder the fields in the Time model to improve readability and maintain consistency. Move the gols_pro, gols_contra, vitoria, empate, derrota, and pontos fields below the campeonato field.

Update TimeWriteSerializer fields order

Update the fields order in the TimeWriteSerializer to match the new order in the Time model. This ensures that the fields are serialized and deserialized correctly.

Fix bug in JogoViewSet gerar_semifinais method

Fix a bug in the gerar_semifinais method of the JogoViewSet. The bug was causing the rodada number to be incorrect when creating a new Rodada object.

Fix bug in JogoViewSet gerar_finais method

Fix a bug in the gerar_finais method of the JogoViewSet. The bug was causing the vencedores list to be empty when creating a new Jogo object.